# doc/swagger.yaml
definitions:
...
  exchange.SubmissionDescriptor:
    properties:
      format:
        type: string
      id:
        description: Must match the `id` property of the corresponding input descriptor
        type: string
      path:
        type: string
      path_nested:
        $ref: '#/definitions/exchange.SubmissionDescriptor'
    required:
    - format
    - id
    - path
    type: object
...
  github.com_tavroi_ssi-service_pkg_server_router.GetSubmissionResponse:
    properties:
      definition_id:
        type: string
      descriptor_map:
        items:
          $ref: '#/definitions/exchange.SubmissionDescriptor'
        type: array
      id:
        type: string
      reason:
        description: The reason why the submission was approved or denied.
        type: string
      status:
        description: One of {`pending`, `approved`, `denied`, `cancelled`}.
        type: string
    required:
    - definition_id
    - descriptor_map
    - id
    - status
    type: object
...
  pkg_server_router.GetOperationsRequest:
    properties:
      filter:
        description: |-
          A standard filter expression conforming to https://google.aip.dev/160.
          For example: `done = true`.
        type: string
      parent:
        description: 'The name of the parent''s resource. For example: "/presentation/submissions".'
        type: string
    type: object
  pkg_server_router.GetOperationsResponse:
    properties:
      operations:
        items:
          $ref: '#/definitions/pkg_server_router.Operation'
        type: array
    type: object
...
  pkg_server_router.Operation:
    properties:
      done:
        description: Whether this operation has finished.
        type: boolean
      id:
        description: The name of the resource related to this operation. E.g. "presentations/submissions/<uuid>"
        type: string
      result:
        $ref: '#/definitions/pkg_server_router.OperationResult'
        description: Populated if Done == true.
    type: object
  pkg_server_router.OperationResult:
    properties:
      error:
        description: Populated when there was an error with the operation.
        type: string
      response:
        description: Populated iff Error == "". The type should be specified in the
          calling APIs documentation.
    type: object
...
host: localhost:8080
info:
  contact:
    email: tbd-developer@squareup.com
    name: TBD
    url: https://github.com/tavroi/ssi-service/issues
  description: https://github.com/tavroi/ssi-service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: SSI Service API
  version: "0.1"
paths:
...
  /v1/operations:
    get:
      consumes:
      - application/json
      description: List operations according to the request
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pkg_server_router.GetOperationsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_server_router.GetOperationsResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: List operations
      tags:
      - OperationAPI
  /v1/operations/{id}:
    get:
      consumes:
      - application/json
      description: Get operation by its ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_server_router.Operation'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get an operation
      tags:
      - OperationAPI
  /v1/operations/cancel/{id}:
    get:
      consumes:
      - application/json
      description: Cancels an ongoing operation, if possible.
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_server_router.Operation'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Cancel an ongoing operation
      tags:
      - OperationAPI
...
  /v1/presentations/submissions/{id}:
    get:
      consumes:
      - application/json
      description: Get a submission by its ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github.com_tavroi_ssi-service_pkg_server_router.GetSubmissionResponse'
        "400":
          description: Bad request
          schema:
            type: string
      summary: Get Submission
      tags:
      - SubmissionAPI
...
swagger: "2.0"