# pkg/service/operation/model.go
package operation

import (
	"github.com/tavroi/ssi-sdk/util"
	"go.einride.tech/aip/filtering"
)

type Result struct {
	Error    string `json:"error,omitempty"`
	Response any    `json:"response,omitempty"`
}

type Operation struct {
	ID     string `json:"json"`
	Done   bool   `json:"done"`
	Result Result `json:"result,omitempty"`
}

type GetOperationsRequest struct {
	Parent string `validate:"required"`
	Filter filtering.Filter
}

func (r GetOperationsRequest) Validate() error {
	return util.NewValidator().Struct(r)
}

type GetOperationsResponse struct {
	Operations []Operation
}

type GetOperationRequest struct {
	ID string `json:"id" validate:"required"`
}

// Validate does struct validation and returns an error when invalid.
func (r GetOperationRequest) Validate() error {
	return util.NewValidator().Struct(r)
}

type CancelOperationRequest struct {
	ID string `json:"id" validate:"required"`
}

// Validate does struct validation and returns an error when invalid.
func (r CancelOperationRequest) Validate() error {
	return util.NewValidator().Struct(r)
}